vault sync mode - rebalancing before claim/withdraw
